requirements.txt
pyyaml
jinja2
tqdm
termcolor
yapsy
gensim
3to2 
language-check